Remove LocationService layer
